When using an object to remove, tries to remove object in array list with same value.  When using int, tries to remove the index from the list.

+++ ralph/java_lib/RalphDataWrappers/ListTypeDataWrapper.java
@@ -112,7 +112,7 @@ public class ListTypeDataWrapper<T,D>
         */
     	if (log_changes && (! incorporating_deltas))
             partner_change_log.add(delete_key_tuple(key_to_delete));
-        val.remove(key_to_delete);
+        val.remove(key_to_delete.intValue());
     }
     
     public void del_key(ActiveEvent active_event,Integer key_to_delete)
